:recycle: [REFACTOR] getting inital staus by hook
